Re-worked the command line tool to support groups of subcommands.
