+++ src/lcmap_client/scripts/cl_tool/main.py
@@ -1,0 +1,5 @@
+from lcmap_client.scripts.cl_tool import (job, lcmap, model, monitor, query,
+                                          system)
+
+def main():
+    lcmap.lcmap()
