+++ src/lcmap_client/scripts/cl_tool/model.py
@@ -1,0 +1,9 @@
+import click
+
+from lcmap_client.scripts.cl_tool.lcmap import lcmap
+
+
+@lcmap.group()
+@click.pass_obj
+def model(config):
+    "TBD: Execute science models in the LCMAP Science Execution Environment."
